pilot/requirements.txt
requests>=2.31.0
pyyaml>=6.0
packaging>=23.0
qdrant-client>=1.9.0
openai>=1.30.0
tiktoken>=0.7.0
GitPython>=3.1.40
pytest>=8.0
python-dotenv>=1.0.0
tenacity>=8.2.3
rich>=13.7.0
